requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
pydantic>=2.4.0
python-dateutil>=2.8.2
numpy>=1.24.0
pandas>=2.0.0
python-multipart>=0.0.6